move typesAttributes function to common
